🚚 #7 Add CalendarManager(function)
