--- ThrowAway/ThrowAway/Calendar/CalendarHelper.swift
@@ -1,26 +1,0 @@
-//
-//  CalendarHelper.swift
-//  ThrowAway
-//
-//  Created by 지준용 on 2022/11/03.
-//
-
-import SwiftUI
-
-class CalendarHelper {
-    let calendar = Calendar.current
-    let dateFormatter = DateFormatter()
-    
-    func monthYearString(_ date: Date) -> String {
-        dateFormatter.dateFormat = "yyyy\("년") LLL"
-        return dateFormatter.string(from: date)
-    }
-    
-    func plusMonth(_ date: Date) -> Date {
-        return calendar.date(byAdding: .month, value: 1, to: date)!
-    }
-    
-    func minusMonth(_ date: Date) -> Date {
-        return calendar.date(byAdding: .month, value: -1, to: date)!
-    }
-}
