+++ ThrowAway/ThrowAway/Calendar/CalendarManager.swift
@@ -1,0 +1,50 @@
+//
+//  CalendarHelper.swift
+//  ThrowAway
+//
+//  Created by 지준용 on 2022/11/03.
+//
+
+import SwiftUI
+
+class CalendarHelper {
+    @EnvironmentObject var dateHolder: DateHolder
+    
+    let calendar = Calendar.current
+    let dateFormatter = DateFormatter()
+    
+    func monthYearString(_ date: Date) -> String {
+        dateFormatter.dateFormat = "yyyy\("년") LLL"
+        return dateFormatter.string(from: date)
+    }
+    
+    // MARK: - DateScroll
+    
+    func plusMonth(_ date: Date) -> Date {
+        return calendar.date(byAdding: .month, value: 1, to: date)!
+    }
+    
+    func minusMonth(_ date: Date) -> Date {
+        return calendar.date(byAdding: .month, value: -1, to: date)!
+    }
+        
+    func daysInMonth(_ date: Date) -> Int {
+        let range = calendar.range(of: .day, in: .month, for: date)!
+        return range.count
+    }
+    
+    func dayOfMonth(_ date: Date) -> Int {
+        let components = calendar.dateComponents([.day], from: date)
+        return components.day!
+    }
+    
+    func firstOfMonth(_ date: Date) -> Date {
+        let components = calendar.dateComponents([.year, .month], from: date)
+        return calendar.date(from: components)!
+    }
+    
+    func weekDay(_ date: Date) -> Int {
+        let components = calendar.dateComponents([.weekday], from: date)
+        return components.weekday! - 1
+    }
+}
